test: separate response-trimmer tests from Medusa shape tests
Extract generic trimToShape() utility tests into their own file
(response-trimmer.test.ts) with additional edge case coverage.
Medusa-specific shape tests remain in medusa-response-shapes.test.ts.
